Reorganized code packages, rearranged copyright statement location, added md file
