--- README.TXT
@@ -1,10 +1,0 @@
-Pokemon IV Breeding (Gen. 6) Probability Calculator program
-Follows the breeding mechanics for Pokemon versions:
-Omega Ruby,
-Alpha Sapphire,
-X,
-Y
-
-(For the Nintendo 3DS)
-
-By Stephen Gung
