+++ README.md
@@ -1,0 +1,9 @@
+My Pokemon IV Breeding (Gen. 6) Probability Calculator program.
+Follows the breeding mechanics for Pokemon versions:
+* Omega Ruby
+* Alpha Sapphire
+* X
+* Y
+(For the Nintendo 3DS)
+
+By Stephen Gung
