+++ org/epifany/pkmnbreedprbltycalc/PkmnCmd.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright 2016, Stephen Gung, All rights reserved
+ */
+
+package org.epifany.pkmnbreedprbltycalc;
+
+/**
+ * @author Stephen Gung
+ */
+public class PkmnCmd{
+	
+	public static final String HP_INC = "hpinc";
+	public static final String HP_EXC = "hpexc";
+	public static final String HP_OPT = "hpopt";
+	
+	public static final String ATK_INC = "atkinc";
+	public static final String ATK_EXC = "atkexc";
+	public static final String ATK_OPT = "atkopt";
+	
+	public static final String DEF_INC = "definc";
+	public static final String DEF_EXC = "defexc";
+	public static final String DEF_OPT = "defopt";
+	
+	public static final String SPA_INC = "spainc";
+	public static final String SPA_EXC = "spaexc";
+	public static final String SPA_OPT = "spaopt";
+	
+	public static final String SPD_INC = "spdinc";
+	public static final String SPD_EXC = "spdexc";
+	public static final String SPD_OPT = "spdopt";
+	
+	public static final String SPE_INC = "speinc";
+	public static final String SPE_EXC = "speexc";
+	public static final String SPE_OPT = "speopt";
+	
+	public static final String PERFECT_0 = "At least 0 perfect stat";
+	public static final String PERFECT_1 = "At least 1 perfect stat";
+	public static final String PERFECT_2 = "At least 2 perfect stats";
+	public static final String PERFECT_3 = "At least 3 perfect stats";
+	public static final String PERFECT_4 = "At least 4 perfect stats";
+	public static final String PERFECT_5 = "At least 5 perfect stats";
+	public static final String PERFECT_6 = "At least 6 perfect stats";
+	
+	public static final String ITEM_OTHER_TEXT = "No Item";
+	public static final String ITEM_POWERBRACER_TEXT = "Power Bracer -- (Atk)";
+	public static final String ITEM_POWERBELT_TEXT = "Power Belt -- (Def)";
+	public static final String ITEM_POWERLENS_TEXT = "Power Lens -- (SpA)";
+	public static final String ITEM_POWERBAND_TEXT = "Power Band -- (SpD)";
+	public static final String ITEM_POWERANKLET_TEXT = "Power Anklet -- (Spe)";
+	public static final String ITEM_POWERWEIGHT_TEXT = "Power Weight -- (HP)";
+	public static final String ITEM_DESTINYKNOT_TEXT = "Destiny Knot";
+	
+}
