+++ org/epifany/pokemon/Gender.java
@@ -1,0 +1,17 @@
+/**
+ * Copyright 2016, Stephen Gung, All rights reserved
+ */
+
+package org.epifany.pokemon;
+
+/**
+ * This class represents the different genders for Pokemon
+ * 
+ * @author Stephen Gung
+ */
+public enum Gender{
+	MALE,
+	FEMALE,
+	// Some Pokemon are genderless
+	GENDERLESS
+}
