--- src/listener/CheckBoxListener.java
@@ -1,30 +1,0 @@
-package listener;
-
-import java.awt.event.ItemEvent;
-import java.awt.event.ItemListener;
-import javax.swing.AbstractButton;
-import pokemoniv_4.PkmnPresenter;
-
-/**
- * Copyright 2016, Stephen Gung, All rights reserved
- * @author Stephen Gung
- */
-public class CheckBoxListener implements ItemListener{
-	private final PkmnPresenter presenter;
-	
-	public CheckBoxListener( PkmnPresenter p){
-		presenter = p;
-	}
-	
-	@Override
-	public void itemStateChanged( ItemEvent ie){
-		AbstractButton button = (AbstractButton)ie.getSource();
-		String command = button.getActionCommand();
-		boolean flag = ( ie.getStateChange() == ItemEvent.SELECTED);
-		System.out.println( command + " " + flag);
-		presenter.updateFlagState( command, flag);
-		presenter.updateFlagStateKey();
-		presenter.updateFlagStateProbability();
-	}
-	
-}
