--- src/pokemon/Gender.java
@@ -1,15 +1,0 @@
-package pokemon;
-
-/**
- * Copyright 2016, Stephen Gung, All rights reserved
- * 
- * This class represents the different genders for Pokemon
- * 
- * @author Stephen Gung
- */
-public enum Gender{
-	MALE,
-	FEMALE,
-	// Some Pokemon are genderless
-	GENDERLESS
-}
